send delete response iff a point was actually deleted
the client sends a barrage of delete requests, and most of the time they
do not delete a point on the actual whiteboard. Server wastes a lot of
overhead to send delete responses that are useless, and clients process
these responses which have no real effect.

Only send delete responses for points that were actually removed from
the whiteboard.